+++ deploy/comfyui_nodes/lush_drive_image/__init__.py
@@ -1,0 +1,10 @@
+from .drive_image_loader import LushLoadImageFromDrive
+
+
+NODE_CLASS_MAPPINGS = {
+    "LushLoadImageFromDrive": LushLoadImageFromDrive,
+}
+
+NODE_DISPLAY_NAME_MAPPINGS = {
+    "LushLoadImageFromDrive": "Lush Media - Load Image from Google Drive",
+}